fix and feat: 1. added redux reset state for the games and 2. home rendering from the sidebar
