+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@expo/html-elements": "^0.3.0",
+    "@react-navigation/drawer": "^6.5.7",
     "@react-navigation/native": "^6.1.2",
     "@react-navigation/native-stack": "^6.9.8",
     "@reduxjs/toolkit": "^1.9.1",
